+++ packages/dataviews/src/components/dataviews-layouts/picker-activity/index.tsx
@@ -267,11 +267,11 @@ export default function ViewPickerActivity< Item >( {
 	const { getSelectionProps } = useSelectionProps( {
 		data: orderedData,
 		getItemId,
+		isItemSelectable: () => true,
 		selection,
 		onChangeSelection,
-		isMultiSelect: isMultiselect,
+		isMultiselect,
 		shouldSelectOnClick: true,
-		isItemSelectable: () => true,
 	} );
 
 	const renderItem = ( item: Item ) => (