+++ packages/dataviews/src/components/dataviews-layouts/picker-grid/index.tsx
@@ -329,11 +329,11 @@ function ViewPickerGrid< Item >( {
 	const { getSelectionProps } = useSelectionProps( {
 		data: orderedData,
 		getItemId,
+		isItemSelectable: () => true,
 		selection,
 		onChangeSelection,
-		isMultiSelect: isMultiselect,
+		isMultiselect,
 		shouldSelectOnClick: true,
-		isItemSelectable: () => true,
 	} );
 
 	const currentPage = view?.page ?? 1;